modules: hal_nordic: Delegate frame Tx to separate thread
Added a Kconfig entry that enables a dedicated Tx thread in the Zephyr
IP stack. This prevents the ieee802154_radio API from being used from
the context of the system workqueue thread, which is undesireable due to
being a blocking API for an undetermined amount of time depending on
usage.

Signed-off-by: Rafał Kuźnia <rafal.kuznia@nordicsemi.no>